Refactor BaseLayer and use Controls component
